+++ tensorqtl/coloc.py
@@ -39,7 +39,7 @@ def coloc(genotypes1_t, genotypes2_t, phenotype1_t, phenotype2_t,
     else:
         r_nominal_t = calculate_corr(
             genotypes1_t, phenotype1_t.reshape(1,-1), residualizer1, return_var=False).squeeze()
-    r2_nominal_t = r_nominal_t.double().pow(2)
+    r2_nominal_t = r_nominal_t.float().pow(2)
 
     if residualizer1 is not None:
         dof = residualizer1.dof
@@ -80,7 +80,7 @@ def coloc(genotypes1_t, genotypes2_t, phenotype1_t, phenotype2_t,
         var_ratio_t = phenotype_var_t.reshape(1,-1) / genotype_var_t.reshape(-1,1)
     else:
         r_nominal_t = calculate_corr(genotypes2_t, phenotype2_t, residualizer2, return_var=False).squeeze()
-    r2_nominal_t = r_nominal_t.double().pow(2)
+    r2_nominal_t = r_nominal_t.float().pow(2)
 
     if residualizer2 is not None:
         dof = residualizer2.dof
