+++ tensorqtl/core.py
@@ -254,7 +254,7 @@ def calculate_interaction_nominal(genotypes_t, phenotypes_t, interaction_t, resi
         b_se_t = torch.sqrt(Xinv[:, torch.eye(nb, dtype=torch.uint8).bool()].unsqueeze(-1).repeat([1,1,nps]) * rss_t.unsqueeze(1).repeat([1,3,1]) / dof)
         # b_se_t = tf.sqrt(tf.tile(tf.expand_dims(tf.matrix_diag_part(Xinv), 2), [1,1,nps]) * tf.tile(tf.expand_dims(rss_t, 1), [1,3,1]) / dof) # (ng x 3) -> (ng x 3 x np)
 
-    tstat_t = (b_t.double() / b_se_t.double()).float()  # (ng x nb x np)
+    tstat_t = (b_t.float() / b_se_t.float()).float()  # (ng x nb x np)
 
     # tdist = tfp.distributions.StudentT(np.float64(dof), loc=np.float64(0.0), scale=np.float64(1.0))
     if not return_sparse:
